+++ app/src/main/java/de/baumann/browser/fragment/Fragment_settings.java
@@ -28,6 +28,8 @@ import de.baumann.browser.Ninja.R;
 
 public class Fragment_settings extends PreferenceFragmentCompat {
 
+    private boolean showContributors;
+
     @Override
     public void onCreatePreferences(Bundle savedInstanceState, String rootKey) {
         setPreferencesFromResource(R.xml.preference_setting, rootKey);
@@ -83,6 +85,7 @@ public class Fragment_settings extends PreferenceFragmentCompat {
         Objects.requireNonNull(findPreference("settings_license")).setOnPreferenceClickListener(new androidx.preference.Preference.OnPreferenceClickListener() {
             @Override
             public boolean onPreferenceClick(androidx.preference.Preference preference) {
+                showContributors = false;
                 showLicenseDialog(getString(R.string.license_title), getString(R.string.license_dialog));
                 return false;
             }
@@ -90,6 +93,7 @@ public class Fragment_settings extends PreferenceFragmentCompat {
         Objects.requireNonNull(findPreference("settings_community")).setOnPreferenceClickListener(new androidx.preference.Preference.OnPreferenceClickListener() {
             @Override
             public boolean onPreferenceClick(androidx.preference.Preference preference) {
+                showContributors = true;
                 showLicenseDialog(getString(R.string.setting_title_community), getString(R.string.cont_dialog));
                 return false;
             }
@@ -97,6 +101,7 @@ public class Fragment_settings extends PreferenceFragmentCompat {
         Objects.requireNonNull(findPreference("settings_license")).setOnPreferenceClickListener(new androidx.preference.Preference.OnPreferenceClickListener() {
             @Override
             public boolean onPreferenceClick(androidx.preference.Preference preference) {
+                showContributors = false;
                 showLicenseDialog(getString(R.string.license_title), getString(R.string.license_dialog));
                 return false;
             }
@@ -104,6 +109,7 @@ public class Fragment_settings extends PreferenceFragmentCompat {
         Objects.requireNonNull(findPreference("settings_info")).setOnPreferenceClickListener(new androidx.preference.Preference.OnPreferenceClickListener() {
             @Override
             public boolean onPreferenceClick(androidx.preference.Preference preference) {
+                showContributors = false;
                 showLicenseDialog(getString(R.string.menu_other_info), getString(R.string.changelog_dialog));
                 return false;
             }
@@ -140,103 +146,118 @@ public class Fragment_settings extends PreferenceFragmentCompat {
         TextView dialog_text = dialogView.findViewById(R.id.dialog_text);
         dialog_text.setText(HelperUnit.textSpannable(text));
 
-        dialog_text.append("\n\nGaukler Faun\n" +
-                "\u25AA Main developer and initiator of this project\n" +
-                "https://github.com/scoute-dich");
+        if (showContributors) {
+            dialog_text.append("\n\nGaukler Faun\n" +
+                    "\u25AA Main developer and initiator of this project\n" +
+                    "https://github.com/scoute-dich");
 
-        dialog_text.append("\n\nAli Demirtas\n" +
-                "\u25AA Turkish Translation\n" +
-                "https://github.com/ali-demirtas");
+            dialog_text.append("\n\nAli Demirtas\n" +
+                    "\u25AA Turkish Translation\n" +
+                    "https://github.com/ali-demirtas");
 
-        dialog_text.append("\n\nCGSLURP LLC\n" +
-                "\u25AA Helped to implement AdBlock and \"request desktop site\" in the previous version of \"FOSS Browser\".\n" +
-                "https://github.com/futrDevelopment");
+            dialog_text.append("\n\nCGSLURP LLC\n" +
+                    "\u25AA Russian translation\n" +
+                    "https://crowdin.com/profile/gaich");
 
-        dialog_text.append("\n\nelement54\n" +
-                "\u25AA fix: keyboard problems (issue #105)&lt;br>\n" +
-                "\u25AA new: option to disable confirmation dialogs on exit\n" +
-                "https://github.com/element54");
+            dialog_text.append("\n\nDmitry Gaich\n" +
+                    "\u25AA Helped to implement AdBlock and \"request desktop site\" in the previous version of \"FOSS Browser\".\n" +
+                    "https://github.com/futrDevelopment");
 
-        dialog_text.append("\n\nelmru\n" +
-                "\u25AA Taiwan Trad. Chinese Translation\n" +
-                "https://github.com/kogiokka");
+            dialog_text.append("\n\nelement54\n" +
+                    "\u25AA fix: keyboard problems (issue #105)\n" +
+                    "\u25AA new: option to disable confirmation dialogs on exit\n" +
+                    "https://github.com/element54");
 
-        /*
+            dialog_text.append("\n\nelmru\n" +
+                    "\u25AA Taiwan Trad. Chinese Translation\n" +
+                    "https://github.com/kogiokka");
 
-        &lt;b>Enrico Monese&lt;/b>&lt;br>
-        \u25AA Italian Translation&lt;br>
-                &lt;i>https://github.com/EnricoMonese&lt;/i>&lt;br>&lt;br>
+            dialog_text.append("\n\nEnrico Monese\n" +
+                    "\u25AA Italian Translation\n" +
+                    "https://github.com/EnricoMonese");
 
-        &lt;b>Francois&lt;/b>&lt;br>
-        \u25AA French Translation&lt;br>
-                &lt;i>https://github.com/franco27&lt;/i>&lt;br>&lt;br>
+            dialog_text.append("\n\nFrancois\n" +
+                    "\u25AA French Translation\n" +
+                    "https://github.com/franco27");
 
-        &lt;b>gh-pmjm&lt;/b>&lt;br>
-        \u25AA Polish translation&lt;br>
-                &lt;i>https://github.com/gh-pmjm&lt;/i>&lt;br>&lt;br>
+            dialog_text.append("\n\ngh-pmjm\n" +
+                    "\u25AA Polish translation\n" +
+                    "https://github.com/gh-pmjm");
 
-        &lt;b>gr1sh&lt;/b>&lt;br>
-        \u25AA fix: some German strings (issues #124, #131)&lt;br>
-                &lt;i>https://github.com/gr1sh&lt;/i>&lt;br>&lt;br>
+            dialog_text.append("\n\ngr1sh\n" +
+                    "\u25AA fix: some German strings (issues #124, #131)\n" +
+                    "https://github.com/gr1sh");
 
-        &lt;b>Harry Heights&lt;/b>&lt;br>
-        \u25AA Documentation of FOSS Browser&lt;br>
-                &lt;i>https://github.com/HarryHeights&lt;/i>&lt;br>&lt;br>
+            dialog_text.append("\n\nHarry Heights\n" +
+                    "\u25AA Documentation of FOSS Browser\n" +
+                    " https://github.com/HarryHeights");
 
-        &lt;b>Heimen Stoffels&lt;/b>&lt;br>
-        \u25AA Dutch translation&lt;br>
-                &lt;i>https://github.com/Vistaus&lt;/i>&lt;br>&lt;br>
+            dialog_text.append("\n\nHeimen Stoffels\n" +
+                    " \u25AA Dutch translation\n" +
+                    "https://github.com/Vistaus");
 
-        &lt;b>Hellohat&lt;/b>&lt;br>
-        \u25AA French translation&lt;br>
-                &lt;i>https://github.com/Hellohat&lt;/i>&lt;br>&lt;br>
+            dialog_text.append("\n\nHellohat\n" +
+                    "\u25AA French translation\n" +
+                    "https://github.com/Hellohat");
 
-        &lt;b>Herman Nunez&lt;/b>&lt;br>
-        \u25AA Spanish translation&lt;br>
-                &lt;i>https://github.com/junior012&lt;/i>&lt;br>&lt;br>
+            dialog_text.append("\n\nHerman Nunez\n" +
+                    "\u25AA Spanish translation\n" +
+                    "https://github.com/junior012");
 
-        &lt;b>Jumping Yang&lt;/b>&lt;br>
-        \u25AA Chinese translation in the previous version of \"FOSS Browser\"&lt;br>
-                &lt;i>https://github.com/JumpingYang001&lt;/i>&lt;br>&lt;br>
+            dialog_text.append("\n\nJumping Yang\n" +
+                    "\u25AA Chinese translation in the previous version of \"FOSS Browser\"\n" +
+                    "https://github.com/JumpingYang001");
 
-        &lt;b>lishoujun&lt;/b>&lt;br>
-        \u25AA Chinese translation&lt;br>
-        \u25AA bug hunting&lt;br>
-                &lt;i>https://github.com/lishoujun&lt;/i>&lt;br>&lt;br>
+            dialog_text.append("\n\nlishoujun\n" +
+                    "\u25AA Chinese translation\n" +
+                    "\u25AA bug hunting\n" +
+                    "https://github.com/lishoujun");
 
-        &lt;b>Peter Bui&lt;/b>&lt;br>
-        \u25AA more font sizes to choose&lt;br>
-                &lt;i>https://github.com/pbui&lt;/i>&lt;br>&lt;br>
+            dialog_text.append("\n\nLukas Novotny\n" +
+                    "\u25AA Czech translation\n" +
+                    "https://crowdin.com/profile/novas78");
 
-        &lt;b>Secangkir Kopi&lt;/b>&lt;br>
-        \u25AA Indonesian translation&lt;br>
-                &lt;i>https://github.com/Secangkir-Kopi&lt;/i>&lt;br>&lt;br>
+            dialog_text.append("\n\nOguz Ersen\n" +
+                    "\u25AA Turkish translation\n" +
+                    "https://crowdin.com/profile/oersen");
 
-        &lt;b>Sérgio Marques&lt;/b>&lt;br>
-        \u25AA Portuguese translation&lt;br>
-                &lt;i>https://github.com/smarquespt&lt;/i>&lt;br>&lt;br>
+            dialog_text.append("\n\nPeter Bui\n" +
+                    "\u25AA more font sizes to choose\n" +
+                    "https://github.com/pbui");
 
-        &lt;b>splinet&lt;/b>&lt;br>
-        \u25AA Russian translation in the previous version of \"FOSS Browser\"&lt;br>
-                &lt;i>https://github.com/splinet&lt;/i>&lt;br>&lt;br>
+            dialog_text.append("\n\nRodolfoCandidoB\n" +
+                    "\u25AA Portuguese, Brazilian translation\n" +
+                    "https://crowdin.com/profile/RodolfoCandidoB");
 
-        &lt;b>SkewedZeppelin&lt;/b>&lt;br>
-        \u25AA Add option to enable Save-Data header&lt;br>
-                &lt;i>https://github.com/SkewedZeppelin&lt;/i>&lt;br>&lt;br>
+            dialog_text.append("\n\nSecangkir Kopi\n" +
+                    "\u25AA Indonesian translation\n" +
+                    "https://github.com/Secangkir-Kopi");
 
-        &lt;b>Tobiplayer3&lt;/b>&lt;br>
-        \u25AA added Qwant search engine&lt;br>
-        \u25AA option to open new tab instead of exiting app&lt;br>
-                &lt;i>https://github.com/Tobiplayer3&lt;/i>&lt;br>&lt;br>
+            dialog_text.append("\n\nSérgio Marques\n" +
+                    "\u25AA Portuguese translation\n" +
+                    "https://github.com/smarquespt");
 
-        &lt;b>Vladimir Kosolapov&lt;/b>&lt;br>
-        \u25AA Russian translation&lt;br>
-                &lt;i>https://github.com/0x264f&lt;/i>&lt;br>&lt;br>
+            dialog_text.append("\n\nsplinet\n" +
+                    "\u25AA Russian translation in the previous version of \"FOSS Browser\"\n" +
+                    "https://github.com/splinet");
 
-        &lt;b>YC L&lt;/b>&lt;br>
-        \u25AA Chinese Translation&lt;br>
-                &lt;i>https://github.com/smallg0at&lt;/i>*/
+            dialog_text.append("\n\nSkewedZeppelin\n" +
+                    "\u25AA Add option to enable Save-Data header\n" +
+                    "https://github.com/SkewedZeppelin");
 
+            dialog_text.append("\n\nTobiplayer\n" +
+                    "\u25AA added Qwant search engine\n" +
+                    "\u25AA option to open new tab instead of exiting\n" +
+                    "https://github.com/Tobiplayer");
+
+            dialog_text.append("\n\nVladimir Kosolapov\n" +
+                    "\u25AA Russian translation\n" +
+                    "https://github.com/0x264f");
+
+            dialog_text.append("\n\nYC L\n" +
+                    "\u25AA Chinese Translation\n" +
+                    "https://github.com/smallg0at");
+        }
 
         dialog_text.setMovementMethod(LinkMovementMethod.getInstance());
         dialog.setContentView(dialogView);