+++ CHANGELOG.md
@@ -1,12 +1,13 @@
 ### v 6.8 (WIP)
 - new: follow system theme
+- new: Amoled theme
 - new: moving to AndroidX libraries
 - new: favorite settings
 - updated: themes
 - updated: Turkish translation
 - updated: Czech translation
 - updated: Brazilian translation
-- updated: Turkish translation
+- updated: Russian translation
 - updated: settings (especially filter settings)
 - updated: saving locations of backups and screenshots
 - removed: tinting of toolbar